moved fire chart from the dashboard column to the actions column
