+++ frontend/src/components/regions/UserActions/index.js
@@ -27,12 +27,17 @@ function UserActionsUnconnected({
   cardSummaries,
   fetchProjectReviewsPages,
   fetchCardCompletions,
+  userBurndownStats,
+  fetchUserBurndownStats,
+
+
   // call logs
   FETCH_RECRUIT_PROJECT_REVIEWS_PAGE,
   FETCH_USER_ACTIONS_CARDS_COMPLETED_PAGE,
 }) {
   let urlParams = useParams() || {};
   const userId = parseInt(urlParams.userId || authedUserId || 0);
+  const currentUserBurndownStats = Object.values(userBurndownStats).filter((snapshot) => snapshot.user === userId);
 
   React.useEffect(() => {
     fetchProjectReviewsPages({
@@ -45,7 +50,9 @@ function UserActionsUnconnected({
 
   React.useEffect(() => {
     fetchCardCompletions({ page: 1, assigneeUserId: userId });
+    fetchUserBurndownStats({ userId });
   }, [fetchCardCompletions, userId]);
+
 
   const latestProjectReviewsCall = getLatestMatchingCall({
     callLog: FETCH_RECRUIT_PROJECT_REVIEWS_PAGE,
@@ -142,6 +149,8 @@ function UserActionsUnconnected({
     actionLogByDate,
     anyLoading,
     handleScroll,
+    currentUserBurndownStats,
+
   };
   return <Presentation {...props} />;
 }
@@ -156,6 +165,8 @@ const mapStateToProps = (state) => {
       state.FETCH_RECRUIT_PROJECT_REVIEWS_PAGE,
     FETCH_USER_ACTIONS_CARDS_COMPLETED_PAGE:
       state.FETCH_USER_ACTIONS_CARDS_COMPLETED_PAGE,
+      userBurndownStats: state.apiEntities.burndownSnapshots || {},
+
   };
 };
 
@@ -168,6 +179,18 @@ const mapDispatchToProps = (dispatch) => {
         )
       );
     },
+        //TODO Implement Page check
+        fetchUserBurndownStats: ({ userId }) => {
+          dispatch(
+            apiReduxApps.FETCH_USER_BURNDOWN_SNAPSHOTS_PAGE.operations.maybeStart({
+              data: { 
+                userId: parseInt(userId),
+                page: 1,
+               },
+            })
+          );
+        },
+    
 
     fetchCardCompletions: ({ assigneeUserId, page }) => {
       dispatch(
